# docs/swagger.yaml
definitions:
  controllers.CarInput:
    properties:
      brand_id:
        type: integer
      description:
        type: string
      is_second:
        type: boolean
      name:
        type: string
      price:
        type: number
      type_id:
        type: integer
    required:
    - brand_id
    - name
    - price
    - type_id
    type: object
  models.BrandCar:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Car:
    properties:
      brand:
        $ref: '#/definitions/models.BrandCar'
      brand_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_second:
        type: boolean
      name:
        type: string
      price:
        type: number
      type:
        $ref: '#/definitions/models.TypeCar'
      type_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.TypeCar:
    properties:
      cars:
        items:
          $ref: '#/definitions/models.Car'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/cms/brand-cars:
    get:
      description: Get a list of all brand cars
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BrandCar'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Get all brand cars
      tags:
      - brand-cars
    post:
      consumes:
      - application/json
      description: Create a new brand car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Brand Car object
        in: body
        name: brand_car
        required: true
        schema:
          $ref: '#/definitions/models.BrandCar'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BrandCar'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Create a new brand car
      tags:
      - brand-cars
  /api/cms/brand-cars/{id}:
    delete:
      description: Delete a specific brand car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Brand Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Delete a brand car
      tags:
      - brand-cars
    get:
      description: Get details of a specific brand car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Brand Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BrandCar'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Get a brand car by ID
      tags:
      - brand-cars
    put:
      consumes:
      - application/json
      description: Update details of a specific brand car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Brand Car ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Brand Car object
        in: body
        name: brand_car
        required: true
        schema:
          $ref: '#/definitions/models.BrandCar'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BrandCar'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Update a brand car
      tags:
      - brand-cars
  /api/cms/cars:
    get:
      description: Get a list of all cars with their types and brands
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Car'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Get all cars
      tags:
      - cars
    post:
      consumes:
      - application/json
      description: Create a new car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car object
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/controllers.CarInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Create a new car
      tags:
      - cars
  /api/cms/cars/{id}:
    delete:
      description: Delete a specific car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Delete a car
      tags:
      - cars
    get:
      description: Get details of a specific car including its type and brand
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Get a car by ID
      tags:
      - cars
    put:
      consumes:
      - application/json
      description: Update details of a specific car
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Car object
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/controllers.CarInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Update a car
      tags:
      - cars
...
swagger: "2.0"